Fix "task_frequency" variable type

+++ src/main/osd/osd.h
@@ -287,7 +287,7 @@ typedef struct osdConfig_s {
     uint8_t logo_on_arming_duration;          // display duration in 0.1s units
     uint8_t camera_frame_width;               // The width of the box for the camera frame element
     uint8_t camera_frame_height;              // The height of the box for the camera frame element
-    uint8_t task_frequency;
+    uint16_t task_frequency;
 } osdConfig_t;
 
 PG_DECLARE(osdConfig_t, osdConfig);
